+++ SomnusLoadBalancer/src/main/java/com/somnus/server/somnuslb/mirrorStatus/MirrorStatusComparator.java
@@ -1,6 +1,7 @@
 package com.somnus.server.somnuslb.mirrorStatus;
 
 import com.somnus.server.somnuslb.mirrorStatus.domain.MirrorInfo;
+import com.somnus.server.somnuslb.mirrorStatus.domain.MirrorState;
 
 import java.util.Comparator;
 
@@ -11,40 +12,27 @@ public class MirrorStatusComparator implements Comparator<MirrorInfo> {
 
     @Override
     public int compare(MirrorInfo o1, MirrorInfo o2) {
+        if (o1.getState() == MirrorState.NORMAL_OPERATION && o2.getState() != MirrorState.NORMAL_OPERATION) {
+            return -1;
+        } else if (o1.getState() != MirrorState.NORMAL_OPERATION && o2.getState() == MirrorState.NORMAL_OPERATION) {
+            return 1;
+        }
+
+        // TODO: check if it maximizes successes and minimizes unwanted actions
         int result = 0;
-        result += normalizeValue(o1.getMirror().getAverageRequestResponseTime(), o2.getMirror().getAverageRequestResponseTime());
-        result += normalizeValue(o1.getMirror().getNumFailedRequests(), o2.getMirror().getNumFailedRequests());
-        result += normalizeValue(o1.getAverageCPUUsage(), o2.getAverageCPUUsage());
-        result += normalizeValue(o1.getAverageMemoryUsage(), o2.getAverageMemoryUsage());
+        result += Long.compare(o1.getMirror().getAverageRequestResponseTime(), o2.getMirror().getAverageRequestResponseTime());
+        result += Integer.compare(o1.getMirror().getNumFailedRequests(), o2.getMirror().getNumFailedRequests());
+        result += Double.compare(o1.getAverageCPUUsage(), o2.getAverageCPUUsage());
+        result += Double.compare(o1.getAverageMemoryUsage(), o2.getAverageMemoryUsage());
+        result += Integer.compare(o1.getNumConsecutiveSuccesses(), o2.getNumConsecutiveSuccesses());
 
-        // changed order on purpose for normalize
-        result += normalizeValue(o2.getNumConsecutiveSuccesses(), o1.getNumConsecutiveSuccesses());
-
-        return result;
-    }
-
-    // TODO: Same method structure, find possible way to reduce code
-
-    public int normalizeValue(int n1, int n2) {
-        if (n1 == 0 && n2 == 0)
-            return 0;
-        else if (n1 == 0 && n2 != 0)
-            return -1;
-        else if (n1 != 0 && n2 == 0)
-            return 1;
-
-        return n1 < n2 ? -n1 / n2 : n2 / n1;
-    }
-
-    public double normalizeValue(double n1, double n2) {
-        if (n1 == 0 && n2 == 0)
-            return 0;
-        else if (n1 == 0 && n2 != 0)
-            return -1;
-        else if(n1 != 0 && n2 == 0)
-            return 1;
-
-        return n1 < n2 ? -n1 / n2 : n2 / n1;
+//        if(result > 0)
+//            return 1;
+//        else if(result < 0)
+//            return -1;
+//
+//        return 0;
+        return result > 0 ? 1 : (result < 0 ? -1 : 0);
     }
 
 }