+++ SomnusLoadBalancer/src/main/java/com/somnus/server/somnuslb/requestHandler/RequestController.java
@@ -1,0 +1,8 @@
+package com.somnus.server.somnuslb.requestHandler;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class RequestController {
+
+}